requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
django-filter
django-crontab
Pillow
